working element-wise matrix-matrix multiply (with caveats), random cleanup
- now have pMat::matrix_elem_mult C = alpha * A .* C
    - element-wise matrix-matrix multiply
    - only works for type double
    - transpose option does not work
    - NOTE: only works with external ScaLAPACK, DOES NOT WORK WITH MKL
- some indentation, comment cleanup

+++ include/misc_tfuncs.hpp
@@ -13,7 +13,7 @@ void printASCIIVecP0(std::string fname, std::vector<T> &Mat, int N)
                 if (typeid(T) == typeid(double))
                         fprintf(fid, "%.9E\n", Mat[i]);
                 if (typeid(T) == typeid(int))
-                        fprintf(fid, "%d\n", Mat[i]);
+                        fprintf(fid, "%d\n", (int)Mat[i]);
         }
         fclose(fid);
 }
